New trove classifier. Better usage string. pep8
The machine I've tested this on has python 2.6 and I can confirm that this
works perfectly well on 2.6 so I added it as a classifier.

The usage string was being set as the description accidentally. By specifying
what we want the description to be, the usage defaults to what it should be
and ``ddg -h`` will be more useful.

I made some changes only to confirm with PEP8. Others (mostly line lenght
issues) are up to @justinls to fix if he so desires.